requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.3
bootstrap4==0.1.0
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.1
cycler==0.12.1
dj-database-url==2.1.0
Django==5.0.6
django-bootstrap4==24.1
django-stdimage==6.0.2
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
filelock==3.16.1
fonttools==4.55.2
gunicorn==21.2.0
idna==3.10
kiwisolver==1.4.7
packaging==24.2
pillow==11.0.0
platformdirs==4.3.6
PyJWT==2.10.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
regex==2024.11.6
requests==2.32.3
six==1.17.0
soupsieve==2.6
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
whitenoise==6.9.0
